tests/wpt/web-platform-tests/scroll-animations/view-timelines/view-timeline-delay.html

<!DOCTYPE html>
<html id="top">
<meta charset="utf-8">
<title>View timeline delay</title>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#viewtimeline-interface">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/web-animations/testcommon.js"></script>
<script src="/scroll-animations/scroll-timelines/testcommon.js"></script>
<script src="/scroll-animations/view-timelines/testcommon.js"></script>
<style>
  #container {
    border:  10px solid lightgray;
    overflow-x: scroll;
    height:  200px;
    width: 200px;
  }
  #content {
    display:  flex;
    flex-flow:  row nowrap;
    justify-content:  flex-start;
    width:  1800px;
    margin: 0;
  }
  .spacer {
    width:  800px;
    display:  inline-block;
  }
  #target {
    background-color:  green;
    height:  100px;
    width:  100px;
    display:  inline-block;
  }
</style>
<body>
  <div id="container">
    <div id="content">
      <div class="spacer"></div>
      <div id="target"></div>
      <div class="spacer"></div>
    </div>
  </div>
</body>
<script type="text/javascript">
  promise_test(async t => {
    // Delays are associated with the animation and not with the timeline.
    // Thus adjusting the delays has no effect on the timeline offsets.  The
    // offsets always correspond to the 'cover' range.
    const verifyTimelineOffsets = anim => {
      const timeline = anim.timeline;
      assert_px_equals(timeline.startOffset, 600, 'startOffset');
      assert_px_equals(timeline.endOffset, 900, 'endOffset');
    };
    await runTimelineDelayTest(t, {
      delay: { phase: 'cover', percent: CSS.percent(0) } ,
      endDelay: { phase: 'cover', percent: CSS.percent(100) },
      rangeStart: 600,
      rangeEnd: 900
    }).then(anim => {
      verifyTimelineOffsets(anim);
    });
    await runTimelineDelayTest(t, {
      delay: { phase: 'contain', percent: CSS.percent(0) } ,
      endDelay: { phase: 'contain', percent: CSS.percent(100) },
      rangeStart: 700,
      rangeEnd: 800
    }).then(anim => {
      verifyTimelineOffsets(anim);
    });
    await runTimelineDelayTest(t, {
      delay: { phase: 'enter', percent: CSS.percent(0) },
      endDelay: { phase: 'enter', percent: CSS.percent(100) },
      rangeStart: 600,
      rangeEnd: 700
    });
    await runTimelineDelayTest(t, {
      delay:  { phase: 'exit', percent: CSS.percent(0) },
      endDelay: { phase: 'exit', percent: CSS.percent(100) },
      rangeStart: 800,
      rangeEnd: 900
    });
    await runTimelineDelayTest(t, {
      delay: { phase: 'contain', percent: CSS.percent(-50) },
      endDelay: { phase: 'enter', percent: CSS.percent(200) },
      rangeStart: 650,
      rangeEnd: 800
    });
    await runTimelineDelayTest(t, {
      delay: { phase: 'enter' },
      endDelay: { phase: 'exit' },
      rangeStart: 600,
      rangeEnd: 900
    });
    await runTimelineDelayTest(t, {
      delay: { percent: CSS.percent(0) },
      endDelay: { percent: CSS.percent(100) },
      rangeStart: 600,
      rangeEnd: 900
    });

  }, 'View timeline with range set via delays.' );
</script>
